Rename pretty_errors->pretyerr
